cleaning(core): move src/core to src/prismic
