+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 export * as Utils from "./utils";
-export * from "./core";
+export * from "./prismic";
 export * as Models from "./models";
 export * as Libraries from "./libraries";
 export * as Mocks from "./libraries/mocks";
